Componentization (#21)
This pull request includes significant updates to the `ProductCard`
component and refactors the routing logic in the application. The
changes aim to enhance the user interface and improve code
maintainability by modularizing components. Below is a summary of the
most important changes:

### Component Refactoring:

*
[`src/app/src/App.svelte`](diffhunk://#diff-4212dfd5a6141ba21a8eff40840b41425f69d4987acc4f079d7c090450621a55L2-R22):
Replaced the existing routing logic with a new `AppRouter` component to
simplify the main application file.
*
[`src/app/src/components/AppRouter.svelte`](diffhunk://#diff-9adcfcdf3c6ed3264975d40a623b49b07bf7798e04b87ca64387d7933d334e30R1-R62):
Introduced the `AppRouter` component to handle routing and navigation
within the application. This component includes logic for catalog
changes and navigation.

### ProductCard Enhancements:

* `src/app/src/components/ProductCard.svelte`: 
* Added `ColorPicker`, `SizePicker`, and `ImageGallery` components to
the `ProductCard` for better product variant selection and image
display.
* Updated the logic to combine images from the product and its variants,
ensuring no duplicate images are displayed.
[[1]](diffhunk://#diff-23f5b15b082d8722af352e544f880da1babef8f2927569a1f05ab79f798db340R71-L59)
[[2]](diffhunk://#diff-23f5b15b082d8722af352e544f880da1babef8f2927569a1f05ab79f798db340L104-R172)
* Implemented functionality to display in-stock and out-of-stock colors
and sizes, improving the user experience when selecting product
variants.
[[1]](diffhunk://#diff-23f5b15b082d8722af352e544f880da1babef8f2927569a1f05ab79f798db340R322-R393)
[[2]](diffhunk://#diff-23f5b15b082d8722af352e544f880da1babef8f2927569a1f05ab79f798db340L184-L188)

### New Components:

*
[`src/app/src/components/ColorPicker.svelte`](diffhunk://#diff-5cbe5f3406c36d70545decf52f533db1eeaae3962b877ea8bc402094ed38c823R1-R37):
Created a `ColorPicker` component to allow users to select product
colors, including handling disabled states for out-of-stock colors.
*
[`src/app/src/components/ImageGallery.svelte`](diffhunk://#diff-af5f410ed1375dec21dd136c50dab9dee1e50d4564d246b36ad9a71778ff3fa1R1-R24):
Created an `ImageGallery` component to display product images in a grid
layout, enhancing the visual presentation of products.